try to not run animation in home
